+++ pakyow-core/lib/pakyow/connection/headers.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "protocol/http/headers"
+
+module Pakyow
+  class Connection
+    class Headers < Protocol::HTTP::Headers
+      def [](key)
+        to_h[key.downcase]
+      end
+    end
+  end
+end